Adding BigDecimal instead of double - stability when dealing with Money

+++ src/main/java/com/taco/challenge/model/FoodPrice.java
@@ -4,11 +4,13 @@ package com.taco.challenge.model;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class FoodPrice {
 
     private int foodId;
     private String foodName;
-    private double foodPrice;
+    private BigDecimal foodPrice;
 }
